Move Context.findActivity extension fun to util package
